agent/requirements.txt
solana==0.30.2
anchorpy==0.18.0
psycopg2-binary==2.9.9
supabase==2.3.0
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
httpx==0.26.0

